+++ Task2/Objects/Task2_RGB_LED_Control.build_log.htm
@@ -22,20 +22,20 @@ Dialog DLL:      TCM.DLL V1.36.5.0
  
 <h2>Project:</h2>
 D:\sprints\GitHub\sprint11\Task2_RGB_LED_Control_V1.0\Task2\Task2_RGB_LED_Control.uvprojx
-Project File Date:  07/02/2023
+Project File Date:  07/04/2023
 
 <h2>Output:</h2>
 *** Using Compiler 'V5.06 update 6 (build 750)', folder: 'C:\Keil_v5\ARM\ARMCC\Bin'
 Rebuild target 'Target 2'
 compiling main.c...
-compiling led.c...
-compiling gpio.c...
 compiling btn.c...
+compiling gpio.c...
+compiling led.c...
 compiling app.c...
 assembling startup_TM4C123.s...
 compiling system_TM4C123.c...
 linking...
-Program Size: Code=2584 RO-data=652 RW-data=24 ZI-data=608  
+Program Size: Code=2632 RO-data=652 RW-data=24 ZI-data=608  
 ".\Objects\Task2_RGB_LED_Control.axf" - 0 Error(s), 0 Warning(s).
 
 <h2>Software Packages used:</h2>
@@ -62,9 +62,9 @@ Package Vendor: Keil
    * Component: ARM::CMSIS:CORE:5.6.0
 
    * Component: Keil::Device:Startup:1.0.1
-      Include file:  Device\Include\TM4C123\TM4C123.h
       Source file:   Device\Source\system_TM4C123.c
       Source file:   Device\Source\ARM\startup_TM4C123.s
+      Include file:  Device\Include\TM4C123\TM4C123.h
 Build Time Elapsed:  00:00:00
 </pre>
 </body>
