+++ Task2/Objects/Task2_RGB_LED_Control.htm
@@ -3,7 +3,7 @@
 <title>Static Call Graph - [.\Objects\Task2_RGB_LED_Control.axf]</title></head>
 <body><HR>
 <H1>Static Call Graph for image .\Objects\Task2_RGB_LED_Control.axf</H1><HR>
-<BR><P>#&#060CALLGRAPH&#062# ARM Linker, 5060750: Last Updated: Sun Jul 02 13:31:13 2023
+<BR><P>#&#060CALLGRAPH&#062# ARM Linker, 5060750: Last Updated: Tue Jul 04 03:27:08 2023
 <BR><P>
 <H3>Maximum Stack Usage =         52 bytes + Unknown(Functions without stacksize, Cycles, Untraceable Function Pointers)</H3><H3>
 Call chain for Maximum Stack Depth:</H3>
@@ -1270,10 +1270,10 @@ Global Symbols
 <P><STRONG><a name="[c2]"></a>__semihosting_library_function</STRONG> (Thumb, 0 bytes, Stack size unknown bytes, indicate_semi.o(.text), UNUSED)
 
 <P><STRONG><a name="[8f]"></a>APP_init</STRONG> (Thumb, 124 bytes, Stack size 16 bytes, app.o(i.APP_init))
-<BR><BR>[Stack]<UL><LI>Max Depth = 44<LI>Call Chain = APP_init &rArr; BUTTON_init &rArr; GPIO_init
+<BR><BR>[Stack]<UL><LI>Max Depth = 44<LI>Call Chain = APP_init &rArr; LED_init &rArr; GPIO_init
 </UL>
-<BR>[Calls]<UL><LI><a href="#[91]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;BUTTON_init
-<LI><a href="#[90]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;LED_init
+<BR>[Calls]<UL><LI><a href="#[90]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;LED_init
+<LI><a href="#[91]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;BUTTON_init
 </UL>
 <BR>[Called By]<UL><LI><a href="#[86]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;main
 </UL>
@@ -1281,14 +1281,14 @@ Global Symbols
 <P><STRONG><a name="[92]"></a>APP_start</STRONG> (Thumb, 332 bytes, Stack size 16 bytes, app.o(i.APP_start))
 <BR><BR>[Stack]<UL><LI>Max Depth = 52<LI>Call Chain = APP_start &rArr; LED_on &rArr; GPIO_write
 </UL>
-<BR>[Calls]<UL><LI><a href="#[93]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;BUTTON_read
-<LI><a href="#[95]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;LED_on
+<BR>[Calls]<UL><LI><a href="#[95]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;LED_on
 <LI><a href="#[94]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;LED_off
+<LI><a href="#[93]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;BUTTON_read
 </UL>
 <BR>[Called By]<UL><LI><a href="#[86]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;main
 </UL>
 
-<P><STRONG><a name="[91]"></a>BUTTON_init</STRONG> (Thumb, 28 bytes, Stack size 16 bytes, btn.o(i.BUTTON_init))
+<P><STRONG><a name="[91]"></a>BUTTON_init</STRONG> (Thumb, 40 bytes, Stack size 16 bytes, btn.o(i.BUTTON_init))
 <BR><BR>[Stack]<UL><LI>Max Depth = 28<LI>Call Chain = BUTTON_init &rArr; GPIO_init
 </UL>
 <BR>[Calls]<UL><LI><a href="#[96]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;GPIO_init
@@ -1307,8 +1307,8 @@ Global Symbols
 <P><STRONG><a name="[96]"></a>GPIO_init</STRONG> (Thumb, 976 bytes, Stack size 12 bytes, gpio.o(i.GPIO_init))
 <BR><BR>[Stack]<UL><LI>Max Depth = 12<LI>Call Chain = GPIO_init
 </UL>
-<BR>[Called By]<UL><LI><a href="#[91]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;BUTTON_init
-<LI><a href="#[90]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;LED_init
+<BR>[Called By]<UL><LI><a href="#[90]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;LED_init
+<LI><a href="#[91]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;BUTTON_init
 </UL>
 
 <P><STRONG><a name="[97]"></a>GPIO_read</STRONG> (Thumb, 58 bytes, Stack size 12 bytes, gpio.o(i.GPIO_read))
@@ -1324,7 +1324,7 @@ Global Symbols
 <LI><a href="#[94]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;LED_off
 </UL>
 
-<P><STRONG><a name="[90]"></a>LED_init</STRONG> (Thumb, 28 bytes, Stack size 16 bytes, led.o(i.LED_init))
+<P><STRONG><a name="[90]"></a>LED_init</STRONG> (Thumb, 40 bytes, Stack size 16 bytes, led.o(i.LED_init))
 <BR><BR>[Stack]<UL><LI>Max Depth = 28<LI>Call Chain = LED_init &rArr; GPIO_init
 </UL>
 <BR>[Calls]<UL><LI><a href="#[96]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;GPIO_init
@@ -1332,7 +1332,7 @@ Global Symbols
 <BR>[Called By]<UL><LI><a href="#[8f]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;APP_init
 </UL>
 
-<P><STRONG><a name="[94]"></a>LED_off</STRONG> (Thumb, 14 bytes, Stack size 16 bytes, led.o(i.LED_off))
+<P><STRONG><a name="[94]"></a>LED_off</STRONG> (Thumb, 26 bytes, Stack size 16 bytes, led.o(i.LED_off))
 <BR><BR>[Stack]<UL><LI>Max Depth = 36<LI>Call Chain = LED_off &rArr; GPIO_write
 </UL>
 <BR>[Calls]<UL><LI><a href="#[98]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;GPIO_write
@@ -1340,7 +1340,7 @@ Global Symbols
 <BR>[Called By]<UL><LI><a href="#[92]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;APP_start
 </UL>
 
-<P><STRONG><a name="[95]"></a>LED_on</STRONG> (Thumb, 14 bytes, Stack size 16 bytes, led.o(i.LED_on))
+<P><STRONG><a name="[95]"></a>LED_on</STRONG> (Thumb, 26 bytes, Stack size 16 bytes, led.o(i.LED_on))
 <BR><BR>[Stack]<UL><LI>Max Depth = 36<LI>Call Chain = LED_on &rArr; GPIO_write
 </UL>
 <BR>[Calls]<UL><LI><a href="#[98]">&gt;&gt;</a>&nbsp;&nbsp;&nbsp;GPIO_write
