+++ Task2/Objects/Task2_RGB_LED_Control.lnp
@@ -1,8 +1,8 @@
 --cpu=Cortex-M4.fp.sp
 ".\objects\gpio.o"
 ".\objects\main.o"
-".\objects\led.o"
 ".\objects\btn.o"
+".\objects\led.o"
 ".\objects\app.o"
 ".\objects\startup_tm4c123.o"
 ".\objects\system_tm4c123.o"
