+++ Task2/Objects/Task2_RGB_LED_Control_Target 2.dep
@@ -1,38 +1,38 @@
 Dependencies for Project 'Task2_RGB_LED_Control', Target 'Target 2': (DO NOT MODIFY !)
 CompilerVersion: 5060750::V5.06 update 6 (build 750)::ARMCC
-F (.\gpio.c)(0x64A15FAF)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\gpio.o --omf_browse .\objects\gpio.crf --depend .\objects\gpio.d)
-I (gpio.h)(0x64A15FAB)
-I (common.h)(0x64A15FAB)
+F (.\MCAL\gpio\gpio.c)(0x64A36EEB)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\gpio.o --omf_browse .\objects\gpio.crf --depend .\objects\gpio.d)
+I (MCAL\gpio\gpio.h)(0x64A36E6E)
+I (MCAL\gpio\../../LIB/common.h)(0x64A36E5E)
 I (C:\Keil_v5\ARM\ARMCC\include\stdint.h)(0x5D9B2678)
-F (.\gpio.h)(0x64A15FAB)()
-F (.\main.c)(0x64A0F3BE)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\main.o --omf_browse .\objects\main.crf --depend .\objects\main.d)
-I (app.h)(0x64A15F0E)
-I (led.h)(0x64A16001)
-I (gpio.h)(0x64A15FAB)
-I (common.h)(0x64A15FAB)
+F (.\MCAL\gpio\gpio.h)(0x64A36E6E)()
+F (.\main.c)(0x64A36E7E)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\main.o --omf_browse .\objects\main.crf --depend .\objects\main.d)
+I (APP/app.h)(0x64A3756C)
+I (APP/../HAL/led/led.h)(0x64A3756C)
+I (APP/../HAL/led/../../MCAL/gpio/gpio.h)(0x64A36E6E)
+I (APP/../HAL/led/../../MCAL/gpio/../../LIB/common.h)(0x64A36E5E)
 I (C:\Keil_v5\ARM\ARMCC\include\stdint.h)(0x5D9B2678)
-I (btn.h)(0x64A15F0E)
-F (.\common.h)(0x64A15FAB)()
-F (.\led.h)(0x64A16001)()
-F (.\led.c)(0x64A15F0E)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\led.o --omf_browse .\objects\led.crf --depend .\objects\led.d)
-I (led.h)(0x64A16001)
-I (gpio.h)(0x64A15FAB)
-I (common.h)(0x64A15FAB)
+I (APP/../HAL/btn/btn.h)(0x64A3756C)
+F (.\LIB\common.h)(0x64A36E5E)()
+F (.\HAL\btn\btn.c)(0x64A3756C)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\btn.o --omf_browse .\objects\btn.crf --depend .\objects\btn.d)
+I (HAL\btn\btn.h)(0x64A3756C)
+I (HAL\btn\../../MCAL/gpio/gpio.h)(0x64A36E6E)
+I (HAL\btn\../../MCAL/gpio/../../LIB/common.h)(0x64A36E5E)
 I (C:\Keil_v5\ARM\ARMCC\include\stdint.h)(0x5D9B2678)
-F (.\btn.h)(0x64A15F0E)()
-F (.\btn.c)(0x64A15F0E)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\btn.o --omf_browse .\objects\btn.crf --depend .\objects\btn.d)
-I (btn.h)(0x64A15F0E)
-I (gpio.h)(0x64A15FAB)
-I (common.h)(0x64A15FAB)
+F (.\HAL\btn\btn.h)(0x64A3756C)()
+F (.\HAL\led\led.c)(0x64A3756C)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\led.o --omf_browse .\objects\led.crf --depend .\objects\led.d)
+I (HAL\led\led.h)(0x64A3756C)
+I (HAL\led\../../MCAL/gpio/gpio.h)(0x64A36E6E)
+I (HAL\led\../../MCAL/gpio/../../LIB/common.h)(0x64A36E5E)
 I (C:\Keil_v5\ARM\ARMCC\include\stdint.h)(0x5D9B2678)
-F (.\app.c)(0x64A15F0E)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\app.o --omf_browse .\objects\app.crf --depend .\objects\app.d)
-I (app.h)(0x64A15F0E)
-I (led.h)(0x64A16001)
-I (gpio.h)(0x64A15FAB)
-I (common.h)(0x64A15FAB)
+F (.\HAL\led\led.h)(0x64A3756C)()
+F (.\APP\app.c)(0x64A3756C)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\app.o --omf_browse .\objects\app.crf --depend .\objects\app.d)
+I (APP\app.h)(0x64A3756C)
+I (APP\../HAL/led/led.h)(0x64A3756C)
+I (APP\../HAL/led/../../MCAL/gpio/gpio.h)(0x64A36E6E)
+I (APP\../HAL/led/../../MCAL/gpio/../../LIB/common.h)(0x64A36E5E)
 I (C:\Keil_v5\ARM\ARMCC\include\stdint.h)(0x5D9B2678)
-I (btn.h)(0x64A15F0E)
-F (.\app.h)(0x64A15F0E)()
+I (APP\../HAL/btn/btn.h)(0x64A3756C)
+F (.\APP\app.h)(0x64A3756C)()
 F (RTE\Device\TM4C123GH6PM\startup_TM4C123.s)(0x64A0F25D)(--cpu Cortex-M4.fp.sp --pd "__EVAL SETA 1" -g --apcs=interwork -I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123--pd "__UVISION_VERSION SETA 529" --pd "_RTE_ SETA 1" --pd "TM4C123GH6PM SETA 1"--list .\listings\startup_tm4c123.lst --xref -o .\objects\startup_tm4c123.o --depend .\objects\startup_tm4c123.d)
 F (RTE\Device\TM4C123GH6PM\system_TM4C123.c)(0x64A0F25D)(--c99 --gnu -c --cpu Cortex-M4.fp.sp -D__EVAL -g -O0 --apcs=interwork --split_sections-I.\RTE\_Target_2-IC:\Users\lenovo\AppData\Local\Arm\Packs\ARM\CMSIS\5.9.0\CMSIS\Core\Include-IC:\Users\lenovo\AppData\Local\Arm\Packs\Keil\TM4C_DFP\1.1.0\Device\Include\TM4C123-D__UVISION_VERSION="529" -D_RTE_ -DTM4C123GH6PM-o .\objects\system_tm4c123.o --omf_browse .\objects\system_tm4c123.crf --depend .\objects\system_tm4c123.d)
 I (C:\Keil_v5\ARM\ARMCC\include\stdint.h)(0x5D9B2678)
