+++ Task2/Objects/app.d
@@ -1,7 +1,7 @@
-.\objects\app.o: app.c
-.\objects\app.o: app.h
-.\objects\app.o: led.h
-.\objects\app.o: gpio.h
-.\objects\app.o: common.h
+.\objects\app.o: APP\app.c
+.\objects\app.o: APP\app.h
+.\objects\app.o: APP\../HAL/led/led.h
+.\objects\app.o: APP\../HAL/led/../../MCAL/gpio/gpio.h
+.\objects\app.o: APP\../HAL/led/../../MCAL/gpio/../../LIB/common.h
 .\objects\app.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdint.h
-.\objects\app.o: btn.h
+.\objects\app.o: APP\../HAL/btn/btn.h
