+++ Task2/Objects/btn.d
@@ -1,5 +1,5 @@
-.\objects\btn.o: btn.c
-.\objects\btn.o: btn.h
-.\objects\btn.o: gpio.h
-.\objects\btn.o: common.h
+.\objects\btn.o: HAL\btn\btn.c
+.\objects\btn.o: HAL\btn\btn.h
+.\objects\btn.o: HAL\btn\../../MCAL/gpio/gpio.h
+.\objects\btn.o: HAL\btn\../../MCAL/gpio/../../LIB/common.h
 .\objects\btn.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdint.h
