+++ Task2/Objects/gpio.d
@@ -1,4 +1,4 @@
-.\objects\gpio.o: gpio.c
-.\objects\gpio.o: gpio.h
-.\objects\gpio.o: common.h
+.\objects\gpio.o: MCAL\gpio\gpio.c
+.\objects\gpio.o: MCAL\gpio\gpio.h
+.\objects\gpio.o: MCAL\gpio\../../LIB/common.h
 .\objects\gpio.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdint.h
