+++ Task2/Objects/led.d
@@ -1,5 +1,5 @@
-.\objects\led.o: led.c
-.\objects\led.o: led.h
-.\objects\led.o: gpio.h
-.\objects\led.o: common.h
+.\objects\led.o: HAL\led\led.c
+.\objects\led.o: HAL\led\led.h
+.\objects\led.o: HAL\led\../../MCAL/gpio/gpio.h
+.\objects\led.o: HAL\led\../../MCAL/gpio/../../LIB/common.h
 .\objects\led.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdint.h
