+++ Task2/Objects/main.d
@@ -1,7 +1,7 @@
 .\objects\main.o: main.c
-.\objects\main.o: app.h
-.\objects\main.o: led.h
-.\objects\main.o: gpio.h
-.\objects\main.o: common.h
+.\objects\main.o: APP/app.h
+.\objects\main.o: APP/../HAL/led/led.h
+.\objects\main.o: APP/../HAL/led/../../MCAL/gpio/gpio.h
+.\objects\main.o: APP/../HAL/led/../../MCAL/gpio/../../LIB/common.h
 .\objects\main.o: C:\Keil_v5\ARM\ARMCC\Bin\..\include\stdint.h
-.\objects\main.o: btn.h
+.\objects\main.o: APP/../HAL/btn/btn.h
