+++ Task2/Task2_RGB_LED_Control.uvprojx
@@ -385,12 +385,12 @@
             <File>
               <FileName>gpio.c</FileName>
               <FileType>1</FileType>
-              <FilePath>.\gpio.c</FilePath>
+              <FilePath>.\MCAL\gpio\gpio.c</FilePath>
             </File>
             <File>
               <FileName>gpio.h</FileName>
               <FileType>5</FileType>
-              <FilePath>.\gpio.h</FilePath>
+              <FilePath>.\MCAL\gpio\gpio.h</FilePath>
             </File>
           </Files>
         </Group>
@@ -410,7 +410,7 @@
             <File>
               <FileName>common.h</FileName>
               <FileType>5</FileType>
-              <FilePath>.\common.h</FilePath>
+              <FilePath>.\LIB\common.h</FilePath>
             </File>
           </Files>
         </Group>
@@ -418,24 +418,24 @@
           <GroupName>HAL</GroupName>
           <Files>
             <File>
-              <FileName>led.h</FileName>
-              <FileType>5</FileType>
-              <FilePath>.\led.h</FilePath>
-            </File>
-            <File>
-              <FileName>led.c</FileName>
+              <FileName>btn.c</FileName>
               <FileType>1</FileType>
-              <FilePath>.\led.c</FilePath>
+              <FilePath>.\HAL\btn\btn.c</FilePath>
             </File>
             <File>
               <FileName>btn.h</FileName>
               <FileType>5</FileType>
-              <FilePath>.\btn.h</FilePath>
+              <FilePath>.\HAL\btn\btn.h</FilePath>
             </File>
             <File>
-              <FileName>btn.c</FileName>
+              <FileName>led.c</FileName>
               <FileType>1</FileType>
-              <FilePath>.\btn.c</FilePath>
+              <FilePath>.\HAL\led\led.c</FilePath>
+            </File>
+            <File>
+              <FileName>led.h</FileName>
+              <FileType>5</FileType>
+              <FilePath>.\HAL\led\led.h</FilePath>
             </File>
           </Files>
         </Group>
@@ -445,12 +445,12 @@
             <File>
               <FileName>app.c</FileName>
               <FileType>1</FileType>
-              <FilePath>.\app.c</FilePath>
+              <FilePath>.\APP\app.c</FilePath>
             </File>
             <File>
               <FileName>app.h</FileName>
               <FileType>5</FileType>
-              <FilePath>.\app.h</FilePath>
+              <FilePath>.\APP\app.h</FilePath>
             </File>
           </Files>
         </Group>
