--- Task2/app.h
@@ -1,18 +1,0 @@
-
-/*
- * app.h
- *
- * Created: 4/9/2023 8:54:47 AM
- *  Author: MOHAMED ABDELSALAM
- */ 
-
-
-#ifndef APP_H_
-#define APP_H_
-#include "led.h"
-#include "btn.h"
-/*APIs*/
-void APP_init(void);
-void APP_start(void);
-
-#endif
